Reset master, checkout head locally
There was a problem where adding new subtree splits for paths that did not exist
when master was originally checked out were not able to be created. Splits were
created correctly based on the commitish ( origin/master, etc. ) but if the path
did not exist in the working directory prior to `git subtree` being called the
subtree command would fail.